feat: Módulo 5 concluído. Função DELETE implementada no front e back end.

+++ pages/api/todos/[id]/index.ts
@@ -9,10 +9,10 @@ export default function handler(
 
     // const todoId = req.query.id;
     // res.end(`Post: ${todoId}`);
-    // if(request.method === "DELETE") {
+    if(request.method === "DELETE") {
             todoController.deleteById(request, response);
             return;
-        // }
+        }
     
         response.status(405).json({
            error: {
